Moved entites of package dk.ule.oapenwb.entity of project oapenwb-javalin into package dk.ule.oapenwb.persistency.entity of project oapenwb-persistency.
