+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/AdminFaces.java
@@ -7,8 +7,8 @@ import com.google.inject.Singleton;
 import com.google.inject.name.Named;
 import dk.ule.oapenwb.entity.content.basedata.*;
 import dk.ule.oapenwb.entity.content.basedata.tlConfig.TypeLanguageConfig;
-import dk.ule.oapenwb.entity.ui.UiLanguage;
-import dk.ule.oapenwb.entity.ui.UiTranslationScope;
+import dk.ule.oapenwb.persistency.entity.ui.UiLanguage;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationScope;
 import dk.ule.oapenwb.faces.UiTranslationSetsFace;
 import dk.ule.oapenwb.faces.admin.*;
 import lombok.Getter;
