+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/DictModule.java
@@ -9,10 +9,10 @@ import com.google.inject.util.Types;
 import dk.ule.oapenwb.base.AppConfig;
 import dk.ule.oapenwb.entity.content.basedata.*;
 import dk.ule.oapenwb.entity.content.basedata.tlConfig.TypeLanguageConfig;
-import dk.ule.oapenwb.entity.ui.UiLanguage;
-import dk.ule.oapenwb.entity.ui.UiTranslation;
-import dk.ule.oapenwb.entity.ui.UiTranslationKey;
-import dk.ule.oapenwb.entity.ui.UiTranslationScope;
+import dk.ule.oapenwb.persistency.entity.ui.UiLanguage;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslation;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationKey;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationScope;
 import dk.ule.oapenwb.faces.*;
 import dk.ule.oapenwb.faces.admin.*;
 import dk.ule.oapenwb.logic.admin.*;
