+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/data/DataInitializer.java
@@ -11,9 +11,9 @@ import dk.ule.oapenwb.base.error.CodeException;
 import dk.ule.oapenwb.entity.basis.RoleType;
 import dk.ule.oapenwb.entity.basis.User;
 import dk.ule.oapenwb.entity.basis.VersionInfo;
-import dk.ule.oapenwb.entity.ui.UiLanguage;
-import dk.ule.oapenwb.entity.ui.UiTranslation;
-import dk.ule.oapenwb.entity.ui.UiTranslationScope;
+import dk.ule.oapenwb.persistency.entity.ui.UiLanguage;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslation;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationScope;
 import dk.ule.oapenwb.util.HibernateConfigurator;
 import dk.ule.oapenwb.util.HibernateUtil;
 import dk.ule.oapenwb.util.Pair;
