+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/data/HibernateConfiguratorImpl.java
@@ -14,10 +14,10 @@ import dk.ule.oapenwb.entity.content.lexemes.lexeme.Tag;
 import dk.ule.oapenwb.entity.content.lexemes.lexeme.Variant;
 import dk.ule.oapenwb.entity.statistics.CondensedData;
 import dk.ule.oapenwb.entity.statistics.SearchRun;
-import dk.ule.oapenwb.entity.ui.UiLanguage;
-import dk.ule.oapenwb.entity.ui.UiResultCategory;
-import dk.ule.oapenwb.entity.ui.UiTranslation;
-import dk.ule.oapenwb.entity.ui.UiTranslationScope;
+import dk.ule.oapenwb.persistency.entity.ui.UiLanguage;
+import dk.ule.oapenwb.persistency.entity.ui.UiResultCategory;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslation;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationScope;
 import dk.ule.oapenwb.util.HibernateConfigurator;
 import lombok.Getter;
 import lombok.Setter;
