+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/data/SaxonDictData.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.data;
 
 import dk.ule.oapenwb.entity.content.basedata.*;
-import dk.ule.oapenwb.entity.ui.UiResultCategory;
+import dk.ule.oapenwb.persistency.entity.ui.UiResultCategory;
 import dk.ule.oapenwb.util.Pair;
 import org.hibernate.Session;
