+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/faces/UiTranslationSetsFace.java
@@ -5,7 +5,7 @@ package dk.ule.oapenwb.faces;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import dk.ule.oapenwb.base.error.CodeException;
-import dk.ule.oapenwb.entity.ui.UiTranslationSet;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationSet;
 import dk.ule.oapenwb.logic.admin.UiTranslationSetsController;
 import dk.ule.oapenwb.util.json.Response;
 import dk.ule.oapenwb.util.json.ResponseStatus;
