+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/faces/admin/UiResultCategoriesFace.java
@@ -4,7 +4,7 @@ package dk.ule.oapenwb.faces.admin;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import dk.ule.oapenwb.entity.ui.UiResultCategory;
+import dk.ule.oapenwb.persistency.entity.ui.UiResultCategory;
 import dk.ule.oapenwb.logic.admin.UiResultCategoriesController;
 
 @Singleton
