+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/admin/UiResultCategoriesController.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.logic.admin;
 
 import com.google.inject.Singleton;
-import dk.ule.oapenwb.entity.ui.UiResultCategory;
+import dk.ule.oapenwb.persistency.entity.ui.UiResultCategory;
 import dk.ule.oapenwb.logic.admin.generic.CEntityController;
 
 @Singleton
