+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/admin/UiTranslationSetsController.java
@@ -8,9 +8,9 @@ import com.google.inject.name.Named;
 import dk.ule.oapenwb.AdminControllers;
 import dk.ule.oapenwb.base.ErrorCode;
 import dk.ule.oapenwb.base.error.CodeException;
-import dk.ule.oapenwb.entity.ui.UiTranslation;
-import dk.ule.oapenwb.entity.ui.UiTranslationKey;
-import dk.ule.oapenwb.entity.ui.UiTranslationSet;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslation;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationKey;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationSet;
 import dk.ule.oapenwb.logic.admin.common.FilterCriterion;
 import dk.ule.oapenwb.logic.admin.common.IRestController;
 import dk.ule.oapenwb.logic.admin.generic.EntityController;
