+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/config/BaseConfig.java
@@ -5,7 +5,7 @@ package dk.ule.oapenwb.logic.config;
 import com.fasterxml.jackson.annotation.JsonView;
 import dk.ule.oapenwb.persistency.entity.Views;
 import dk.ule.oapenwb.entity.content.basedata.*;
-import dk.ule.oapenwb.entity.ui.UiLanguage;
+import dk.ule.oapenwb.persistency.entity.ui.UiLanguage;
 import lombok.Data;
 
 import java.util.LinkedList;
