+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/config/Config.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.logic.config;
 
 import dk.ule.oapenwb.entity.content.basedata.*;
-import dk.ule.oapenwb.entity.ui.UiLanguage;
+import dk.ule.oapenwb.persistency.entity.ui.UiLanguage;
 import lombok.Getter;
 
 import java.util.List;
