+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/l10n/L10nController.java
@@ -6,9 +6,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.inject.Singleton;
 import dk.ule.oapenwb.base.ErrorCode;
 import dk.ule.oapenwb.base.error.CodeException;
-import dk.ule.oapenwb.entity.ui.UiLanguage;
-import dk.ule.oapenwb.entity.ui.UiTranslation;
-import dk.ule.oapenwb.entity.ui.UiTranslationScope;
+import dk.ule.oapenwb.persistency.entity.ui.UiLanguage;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslation;
+import dk.ule.oapenwb.persistency.entity.ui.UiTranslationScope;
 import dk.ule.oapenwb.util.HibernateUtil;
 import dk.ule.oapenwb.util.TimeUtil;
 import io.javalin.json.JavalinJackson;
